# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Todo List",
        "description": "Dokumentasi api yang tersedia",
        "contact": {
            "email": "iqbal.ramadhani@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User Profile",
                "description": "Show detail user account",
                "operationId": "dd48b9887ac0e98fac70ebd3da3eb645",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "user_profile": {
                                        "id": 123,
                                        "name": "tes",
                                        "email": "iqbal2@email.com",
                                        "created_at": "15-03-2024 06:06:30",
                                        "updated_at": "15-03-2024 06:06:30",
                                        "type": "organisasi",
                                        "no_handphone": null,
                                        "kode_refferal": null,
                                        "no_rekening": null,
                                        "bank": null,
                                        "atas_nama_bank": null,
                                        "url_photo": null,
                                        "wakaf_maksimal": 0,
                                        "bulan_maksimal_pembayaran": 0
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error Access Token",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/update-profile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update Profile",
                "description": "Update user profile data",
                "operationId": "1caf4d5e3c235a7e4632213c819dbbec",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "no_handphone": {
                                        "description": "No Handphone",
                                        "type": "string"
                                    },
                                    "no_rekening": {
                                        "description": "No Rekening",
                                        "type": "string"
                                    },
                                    "nama_bank": {
                                        "description": "Nama Bank",
                                        "type": "string"
                                    },
                                    "pemilik_rekening": {
                                        "description": "Pemilik Rekening",
                                        "type": "string"
                                    },
                                    "wakaf_maksimal": {
                                        "description": "Wakaf maksimal",
                                        "type": "number"
                                    },
                                    "bulan_maksimal_pembayaran": {
                                        "description": "Perbulan maksimal pembayaran",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "user_data": {
                                        "id": "integer",
                                        "name": "string",
                                        "email": "string",
                                        "role": "string",
                                        "no_handphone": "string",
                                        "kode_refferal": "string",
                                        "no_rekening": "string",
                                        "bank": "string",
                                        "atas_nama_bank": "string",
                                        "url_photo": "string",
                                        "wakaf_maksimal": "integer",
                                        "bulan_maksimal_pembayaran": "integer",
                                        "organisasi": "string",
                                        "profile_lengkap": "boolean"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation if photo profile still empty after update data",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "user_data": {
                                        "id": "integer",
                                        "name": "string",
                                        "email": "string",
                                        "role": "string",
                                        "no_handphone": "string",
                                        "kode_refferal": "string",
                                        "no_rekening": "string",
                                        "bank": "string",
                                        "atas_nama_bank": "string",
                                        "url_photo": "string",
                                        "wakaf_maksimal": "integer",
                                        "bulan_maksimal_pembayaran": "integer",
                                        "organisasi": "string",
                                        "profile_lengkap": "boolean"
                                    },
                                    "message": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "error": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}